+++ src/main/java/org/example/final_project/services/UserService.java
@@ -11,27 +11,27 @@ public interface UserService {
 	
 	public User registerUser(User user) throws UserException;
 	
-	public User findUserById(Integer userId) throws UserException;
+	public User findUserById(Long userId) throws UserException;
 	
 	public User findUserProfile(String token) throws UserException;
 	
 	public User findUserByUsername(String username) throws UserException;
 	
-	public String followUser(Integer reqUserId,Integer followUserId) throws UserException;
+	public String followUser(Long reqUserId,Long followUserId) throws UserException;
 	
-	public String unfollowUser(Integer reqUserId, Integer unfollowUserId) throws UserException;
+	public String unfollowUser(Long reqUserId, Long unfollowUserId) throws UserException;
 
-	public String removeFollower(Integer reqUserId, Integer followerUserId) throws UserException;
+	public String removeFollower(Long reqUserId, Long followerUserId) throws UserException;
 
-	public List<User> findUsersByUserIds(List<Integer> userIds);
+	public List<User> findUsersByUserIds(List<Long> userIds);
 	
 	public List<User> searchUser(String query) throws UserException;
 	
 	public List<User> popularUser();
 
-	public List<User> getFollowingUsers(Integer userId);
+	public List<User> getFollowingUsers(Long userId);
 
-	public List<User> getFollowerUsers(Integer userId);
+	public List<User> getFollowerUsers(Long userId);
 
 	public User updateUserDetails(User updatedUser, User existingUser) throws UserException;
